+++ src/backend/demo/management/commands/create_demo.py
@@ -174,7 +174,8 @@ def create_demo(stdout):
                     is_superuser=False,
                     is_active=True,
                     is_staff=False,
-                    language=random.choice(settings.LANGUAGES)[0],
+                    language=dev_user["language"]
+                    or random.choice(settings.LANGUAGES)[0],
                 )
             )