+++ src/frontend/apps/impress/src/core/auth/api/types.ts
@@ -4,10 +4,12 @@
  * @property {string} id - The id of the user.
  * @property {string} email - The email of the user.
  * @property {string} name - The name of the user.
+ * @property {string} language - The language of the user.
  */
 export interface User {
   id: string;
   email: string;
   full_name: string;
   short_name: string;
+  language: string;
 }