+++ src/frontend/apps/impress/src/i18n/initI18n.ts
@@ -2,7 +2,6 @@ import i18n from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import { initReactI18next } from 'react-i18next';
 
-import { BASE_LANGUAGE, LANGUAGES_ALLOWED, LANGUAGE_COOKIE_NAME } from './conf';
 import resources from './translations.json';
 
 i18n
@@ -10,18 +9,20 @@ i18n
   .use(initReactI18next)
   .init({
     resources,
-    fallbackLng: BASE_LANGUAGE,
-    supportedLngs: Object.keys(LANGUAGES_ALLOWED),
     detection: {
       order: ['cookie', 'navigator'], // detection order
       caches: ['cookie'], // Use cookies to store the language preference
-      lookupCookie: LANGUAGE_COOKIE_NAME,
+      lookupCookie: 'docs_language',
       cookieMinutes: 525600, // Expires after one year
+      cookieOptions: {
+        sameSite: 'lax',
+      },
     },
     interpolation: {
       escapeValue: false,
     },
-    preload: Object.keys(LANGUAGES_ALLOWED),
+    preload: Object.keys(resources),
+    lowerCaseLng: true,
     nsSeparator: false,
     keySeparator: false,
   })